+++ webpack.config.babel.js
@@ -1,25 +1,32 @@
 import webpack from 'webpack';
 import path from 'path';
+import styleLintPlugin from 'stylelint-webpack-plugin';
 
 export default {
 	devtool: 'eval',
 	context: path.resolve(__dirname + '/src'),
 	entry: {
-	  app: './app.jsx'
+		app: './app.jsx'
 	},
 	output: {
-	    path: path.resolve(__dirname + '/dist'),
-	    filename: '[name].bundle.js'
+		path: path.resolve(__dirname + '/dist'),
+		filename: '[name].bundle.js'
 	},
 	plugins: [
-	  new webpack.DefinePlugin({
-	    'process.env': {
-	      'NODE_ENV': JSON.stringify('production')
-	     }
-	  })
+		new styleLintPlugin({
+			configFile: './node_modules/stylelint-config-standard/index.js',
+			files: '**/*.css',
+			failOnError: false
+		}),
+		new webpack.DefinePlugin({
+			'process.env': {
+				'NODE_ENV': JSON.stringify('production')
+			}
+		})
 	],
 	resolve: {
-	   extensions: ['', '.js', '.jsx']
+		modulesDirectories: ['src', 'node_modules'],
+		extensions: ['', '.js', '.jsx']
 	},
 	module: {
 		preLoaders: [
@@ -32,7 +39,7 @@ export default {
 		loaders: [
 			{
 				test: /\.css$/,
-				loader: 'style!css',
+				loader: 'style!css?modules&importLoaders=1&localIdentName=[name]__[local]___[hash:base64:5]',
 				exclude: /node_modules/
 			},
 			{